+++ facebook/__init__.py
@@ -47,7 +47,6 @@ FACEBOOK_GRAPH_URL = "https://graph.facebook.com/"
 FACEBOOK_WWW_URL = "https://www.facebook.com/"
 FACEBOOK_OAUTH_DIALOG_PATH = "dialog/oauth?"
 VALID_API_VERSIONS = [
-    "2.8",
     "2.9",
     "2.10",
     "2.11",